requirements.txt
# Core Framework
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# AI/ML Libraries
langchain==0.3.14
langchain-community==0.3.14
langchain-huggingface==0.1.2
transformers==4.47.1
torch==2.5.1
sentence-transformers==3.3.1

# Vector Store
faiss-cpu==1.9.0

# Document Processing
pypdf==4.1.0

# API Clients
huggingface-hub>=0.23.0
requests==2.31.0

# Optimization
pulp==2.9.0

# Retry Logic & Circuit Breaker
tenacity==9.0.0
pybreaker==1.2.0

# Input Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# API Integration
together>=1.5.0
httpx>=0.27.0

# Monitoring & Logging
python-multipart==0.0.20

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-mock>=3.14.0
pytest-asyncio>=0.24.0
